final pathfinding fixes

+++ backend/Dijkstra/RoutePreview.cs
@@ -7,7 +7,7 @@ namespace PublicTransportNavigator.Dijkstra
         public string Id { get; set; }
         public TimeSpan DepartureTime { get; set; }
         public TimeSpan DestinationTime { get; set; }
-        public int TravelTime { get; set; }
+        public double TravelTime { get; set; }
         public List<string> BusNumbers { get; set; } = [];
         public List<Coordinate> Coordinates { get; set; } = [];
     }
